refactor(backend): move ContactListAPI to external apis
